refactor auth routes
